do some restructuring optimising framework integration
