+++ eth_vertigo/interfaces/truffle/__init__.py
@@ -1,0 +1,1 @@
+from eth_vertigo.interfaces.truffle.campaign import TruffleCampaign
