--- eth_vertigo/test_runner/truffle/__init__.py
@@ -1,3 +1,0 @@
-from eth_vertigo.test_runner.truffle.truffle_runner import TruffleRunner
-from eth_vertigo.test_runner.truffle.truffle_runner_factory import TruffleRunnerFactory
-from eth_vertigo.test_runner.truffle.truffle_tester import TruffleTester
