+++ test/test_runner/truffle/test_truffle_tester.py
@@ -1,4 +1,4 @@
-from eth_vertigo.test_runner.truffle.truffle_tester import TruffleTester
+from eth_vertigo.interfaces.truffle.tester import TruffleTester
 import pytest
 
 
@@ -8,9 +8,8 @@ def test_run_test_command():
 
     # Act and Assert
     with pytest.raises(NotImplementedError):
-        tester.run_test_command(
+        tester._run_test_command(
             working_directory="/",
-            file_name=None,
             network_name=None,
             timeout=10
         )
